pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pylrclib"
version = "1.1.0"
description = "A modern multi-command CLI for inspecting, cleansing, and uploading lyrics to LRCLIB."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Harmonese"}]
license = {text = "MIT"}
keywords = ["lrclib", "lyrics", "lrc", "music", "cli"]
dependencies = [
  "requests>=2.31.0",
  "mutagen>=1.47.0",
  "PyYAML>=6.0.0"
]

[tool.setuptools.packages.find]
include = ["pylrclib*"]

[project.scripts]
pylrclib = "pylrclib.cli.main:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
